% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cop_data.R
\docType{data}
\name{mzXML_data}
\alias{mzXML_data}
\title{mzXML_data.}
\format{
A object of class mzXML. A test dataset imported by read.mzXML() from a GC-APCI measurement on a Bruker impact II, exported by Compass to an mzXML file.
}
\source{
Jan Lisec (jan.lisec@bam.de)
}
\usage{
data(mzXML_data)
}
\description{
mzXML_data.
}
\keyword{datasets}
