% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cop_utils.R
\name{verify_suggested}
\alias{verify_suggested}
\title{verify_suggested.}
\usage{
verify_suggested(pkg)
}
\arguments{
\item{pkg}{Package names to be checked.}
}
\value{
NULL.
}
\description{
Check if packages are available and stop function otherwise.
}
\keyword{internal}
