% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.mzXML.R
\name{write.mzXML}
\alias{write.mzXML}
\title{Write a standard mzXML.}
\usage{
write.mzXML(mzXML, filename, precision = c("32", "64"))
}
\arguments{
\item{mzXML}{The generic mzXML object}

\item{filename}{A mzXML filename to export to.}

\item{precision}{Either '32' or '64' byte.}
}
\value{
Nothing. mzXML object is exported to filename. If it was imported using \code{\link{read.mzXML}} it should contain all previous fields.
}
\description{
\code{write.mzXML} is a copy from the caMassClass package which is no longer actively on CRAN
}
\details{
The main task of \code{\link{read.mzXML}} and write.mzXML functions is to extract and save scan data of mzXML files. 
In addition attempt is made to keep all other sections of mzXML file as unparsed XML code, so the data can be extracted latter or saved into new mzXML files.
Those unparsed sections are stored as XML text.
}

