\name{xcmsRaw_data}
\alias{xcmsRaw_data}
\docType{data}
\title{A test dataset of class \code{xcmsRaw}.}
\description{A few seconds extracted from a GC-APCI-MS run using a Bruker impact II.}
\usage{data("xcmsRaw_data")}
\format{
  The format is based on class 'xcmsRaw' [package "xcms"] with several slots. The slots which
  will be used by the functions are 'env', 'scantime' and 'scanindex'. Only individual intensity
  values within 'env$intensity' will be modified.
}
\keyword{datasets}